+++ src/main/java/com/example/book/service/BookService.java
@@ -16,7 +16,7 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    public void deleteBook(int id) {
+    public void deleteBook(Long id) {
         bookRepository.deleteById(id);
     }
